Move Provider to internal package.
Provider is shared by all of the instrumentation packages, but it isn't part of
the API.
