+++ core/src/main/java/com/google/instrumentation/stats/Stats.java
@@ -13,7 +13,7 @@
 
 package com.google.instrumentation.stats;
 
-import com.google.instrumentation.common.Provider;
+import com.google.instrumentation.internal.Provider;
 import javax.annotation.Nullable;
 
 /**
