+++ core/src/main/java/com/google/instrumentation/trace/Tracing.java
@@ -14,7 +14,7 @@
 package com.google.instrumentation.trace;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.google.instrumentation.common.Provider;
+import com.google.instrumentation.internal.Provider;
 import java.util.logging.Level;
 import java.util.logging.Logger;
